+++ packages/ui/src/index.ts
@@ -45,3 +45,4 @@ export * from './components/TableFilters';
 export * from './components/ChipPicker';
 export * from './components/SearchInput';
 export * from './components/SegmentedControl';
+export * from './components/AlertBanner';